Move domain::Tags::InitialRefinementLevels to DomainCreator library
