+++ src/ParallelAlgorithms/LinearSolver/Multigrid/Tags.hpp
@@ -13,7 +13,7 @@
 #include "DataStructures/DataBox/PrefixHelpers.hpp"
 #include "DataStructures/DataBox/Tag.hpp"
 #include "DataStructures/Variables.hpp"
-#include "Domain/Tags.hpp"
+#include "Domain/Creators/Tags/InitialRefinementLevels.hpp"
 #include "NumericalAlgorithms/Spectral/Mesh.hpp"
 #include "Options/Auto.hpp"
 #include "Options/Options.hpp"
